backend/requirements.txt
# ================================
# XBoard Modern 后端依赖包
# ================================

# Web框架
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0

# 数据库
sqlalchemy==2.0.23
alembic==1.12.1

# 数据库驱动
pymysql==1.1.0
psycopg2-binary==2.9.9
aiosqlite==0.19.0

# 认证和安全
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# 邮件
aiosmtplib==3.0.1
email-validator==2.1.0

# 缓存
redis==5.0.1
aioredis==2.0.1

# 数据处理
pydantic==2.5.0
pydantic-settings==2.1.0

# HTTP客户端
httpx==0.25.2
aiohttp==3.9.1

# 工具库
python-dotenv==1.0.0
python-dateutil==2.8.2
pytz==2023.3

# 日志
loguru==0.7.2

# 文件处理
aiofiles==23.2.1
Pillow==10.1.0

# 支付相关
cryptography==41.0.8
pycryptodome==3.19.0

# 二维码
qrcode[pil]==7.4.2

# 配置管理
pyyaml==6.0.1

# 开发工具
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
isort==5.12.0

# 监控和性能
prometheus-client==0.19.0 